requirements.txt
arrow==1.3.0
astral==3.2
attrs==25.3.0
azure-common==1.1.28
azure-core==1.30.1
azure-identity==1.17.1
azure-mgmt-apimanagement==4.0.1
azure-mgmt-core==1.4.0
azure-mgmt-network==28.1.0
azure-mgmt-subscription==3.1.1
azure-storage-blob==12.19.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
easygui==0.98.3
ics==0.7.2
idna==3.4
isodate==0.6.1
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
oauthlib==3.2.2
portalocker==2.10.1
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.9.0
PyMuPDF==1.23.4
PyMuPDFb==1.23.3
pyOpenSSL==23.2.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
pytz==2025.2
pywin32==306
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
TatSu==5.13.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.10.0
tzdata==2025.2
urllib3==2.0.4
wget==3.2
